Add handling of empty YAML in codecs.load_all_yaml
Updates load_all_yaml() to handle empty documents in the yaml stream.  For example, load_all_yaml("---\n---") would previously raise with an TypeError as it tried to iterate on the keys in the empty (but valid) YAML document that resolved to a None object.  This update skips any empty (None) objects.

Also adds tests for this case.