feat: add client.apply_many()
This adds an apply_many function, similar in intent to `kubectl apply -f some-yaml.yaml`, that applies a collection of Resources.  This function also applies the Resources in a safe order by kind, in order to avoid cases like trying to create an instance of a Custom Resource before applying the CustomResourceDefinition.