feat: add client.delete_many()
This adds a delete_many() function, similar to apply_many() and in intent to `kubectl delete -f some-yaml.yaml`, that deletes a collection of Resources.  This function deletes the resources in a safe order by kind, in order to avoid cases like trying to delete an instance of a Custom Resource after the CustomResourceDefinition has already been deleted.

Also added here are tests for this function.