feat: add sort_objects
This adds the `sort_objects()` helper to sort iterables of resource objects in particular orders.  Currently supported is the `kind` order, which orders resources by their kind in an apply-friendly order (avoiding things like applying an instance of a CR before applying the CRD, etc.).

Also supported is an optional `reverse` parameter in the sorting function to reverse the returned order.