fix: remove chance for duplication of CRD between tests
Previously, test_load_in_cluster_generic_resources used a CRD for CronTab that was shared with another test.  If that CRD was applied and not deleted in a previous test, test_load_in_cluster_generic_resources fails because it asserts the CRD does not exist at the start of the test.  This change makes a new (very similar) CRD so we cannot have this collision.